requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2

pillow==10.4.0
opencv-python==4.10.0.84
numpy==2.0.2

pytesseract==0.3.13
paddleocr==2.7.3
paddlepaddle==2.6.2

layoutparser==0.3.4
detectron2==0.6

python-multipart==0.0.9
PyYAML==6.0.2

mlflow==2.16.2
rapidfuzz==3.9.7

openai==1.44.1